+++ nevergrad/optimization/test_optimizerlib.py
@@ -63,6 +63,7 @@ def check_optimizer(optimizer_cls: Union[base.OptimizerFamily, Type[base.Optimiz
     assert (optimizer.current_bests["pessimistic"].pessimistic_confidence_bound ==
             min(v.pessimistic_confidence_bound for v in archive.values()))
     # add a random point to test tell_not_asked
+    assert not optimizer._asked, "All `ask`s  should have been followed by a `tell`"
     try:
         candidate = optimizer.create_candidate.from_data(np.random.normal(0, 1, size=optimizer.dimension))
         optimizer.tell_not_asked(candidate, 12.)
@@ -72,6 +73,7 @@ def check_optimizer(optimizer_cls: Union[base.OptimizerFamily, Type[base.Optimiz
                                  "at tell_not_asked if they do not support it") from e
     else:
         assert optimizer.num_tell == budget + 1
+        assert optimizer.num_tell_not_asked == 1
 
 
 SLOW = ["NoisyDE", "NoisyBandit", "SPSA", "NoisyOnePlusOne", "OptimisticNoisyOnePlusOne", "ASCMADEthird", "ASCMA2PDEthird", "MultiScaleCMA",